feat: 客户详情 + review 修改

+++ yudao-module-crm/yudao-module-crm-biz/src/main/java/cn/iocoder/yudao/module/crm/controller/admin/customer/vo/CrmCustomerBaseVO.java
@@ -68,7 +68,7 @@ public class CrmCustomerBaseVO {
     private String remark;
 
     @Schema(description = "地区编号", example = "20158")
-    private Long areaId;
+    private Integer areaId;
 
     @Schema(description = "详细地址", example = "北京市海淀区")
     private String detailAddress;
